Server configuration is read from the config file
