+++ routeguide/src/main/scala/runtime/server.scala
@@ -17,116 +17,27 @@
 package routeguide
 package runtime
 
-import java.util.concurrent.atomic.AtomicReference
-
 import cats.{~>, Comonad}
+import freestyle.rpc.server._
 import routeguide.protocols._
-import routeguide.codecs._
-import monix.reactive.Observable
+import routeguide.runtime.handlers.RouteGuideServiceHandler
 
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext, Future}
-import scala.io.Source
+import scala.util.{Failure, Success, Try}
 
-object server extends CommonImplicits {
+object server {
 
-  import io.circe.parser.decode
+  import common._
 
-  val features: List[Feature] =
-    decode[FeatureDatabase](
-      Source
-        .fromInputStream(getClass.getClassLoader.getResourceAsStream("route_guide_db.json"))
-        .mkString) match {
-      case Right(fList) => fList.feature
-      case Left(e) =>
-        println(s"Decoding failure: $e")
-        throw e
-    }
+  trait Instances {
 
-  class RouteGuideHandler extends RouteGuideService.Handler[Future] {
+    private[this] val atMostDuration: FiniteDuration = 5.seconds
 
-    // AtomicReference as an alternative to ConcurrentMap<Point, List<RouteNote>>?
-    private val routeNotes: AtomicReference[Map[Point, List[RouteNote]]] =
-      new AtomicReference[Map[Point, List[RouteNote]]](Map.empty)
-
-    override protected[this] def getFeature(point: protocols.Point): Future[Feature] =
-      Future.successful(point.findFeatureIn(features))
-
-    override protected[this] def listFeatures(
-        rectangle: protocols.Rectangle): Future[Observable[Feature]] = {
-      val left   = Math.min(rectangle.lo.longitude, rectangle.hi.longitude)
-      val right  = Math.max(rectangle.lo.longitude, rectangle.hi.longitude)
-      val top    = Math.max(rectangle.lo.latitude, rectangle.hi.latitude)
-      val bottom = Math.min(rectangle.lo.latitude, rectangle.hi.latitude)
-
-      val observable = Observable.fromIterable(
-        features.filter { feature =>
-          val lat = feature.location.latitude
-          val lon = feature.location.longitude
-          feature.valid && lon >= left && lon <= right && lat >= bottom && lat <= top
-
-        }
-      )
-
-      Future.successful(observable)
-    }
-
-    override protected[this] def recordRoute(
-        points: Observable[protocols.Point]): Future[RouteSummary] = {
-      points
-        .foldLeftL((RouteSummary(0, 0, 0, 0), None: Option[Point], System.nanoTime())) {
-          case ((routeSummary, previous, startTime), point) =>
-            // For each point after the first, add the incremental distance from the previous point to
-            // the total distance value.
-            val counter  = if (point.findFeatureIn(features).valid) 1 else 0
-            val distance = previous.map(calcDistance(_, point)) getOrElse 0
-            val updatedRouteSummary: RouteSummary = routeSummary.copy(
-              point_count = routeSummary.point_count + 1,
-              feature_count = routeSummary.feature_count + counter,
-              distance = routeSummary.distance + distance,
-              elapsed_time = NANOSECONDS.toSeconds(System.nanoTime() - startTime).toInt
-            )
-            (updatedRouteSummary, Some(point), startTime)
-        }
-        .map(_._1)
-        .runAsync
-    }
-
-    override protected[this] def routeChat(
-        routeNotes: Observable[protocols.RouteNote]): Future[Observable[RouteNote]] =
-      Future.successful {
-        routeNotes flatMap { note =>
-          addNote(note)
-          Observable.fromIterable(getOrCreateNotes(note.location))
-        }
-      }
-
-    private[this] def addNote(note: RouteNote): Unit =
-      routeNotes.updateAndGet { notes: Map[Point, List[RouteNote]] =>
-        val newRouteNotes = notes.getOrElse(note.location, Nil) :+ note
-        notes + (note.location -> newRouteNotes)
-      }
-
-    private[this] def getOrCreateNotes(point: Point): List[RouteNote] =
-      routeNotes.get.getOrElse(point, Nil)
-  }
-
-  trait Implicits {
-
-    import cats.implicits._
-    import freestyle.rpc.server._
-    import freestyle.rpc.server.handlers._
-    import freestyle.rpc.server.implicits._
-    import freestyle.rpc.client.implicits._
-
-    implicit val finiteDuration: FiniteDuration = 5.seconds
-
-    implicit def futureComonad(
-        implicit ec: ExecutionContext,
-        atMost: FiniteDuration): Comonad[Future] =
+    implicit def futureComonad(implicit ec: ExecutionContext): Comonad[Future] =
       new Comonad[Future] {
         def extract[A](x: Future[A]): A =
-          Await.result(x, atMost)
+          Await.result(x, atMostDuration)
 
         override def coflatMap[A, B](fa: Future[A])(f: (Future[A]) => B): Future[B] = Future(f(fa))
 
@@ -134,14 +45,36 @@ object server extends CommonImplicits {
           fa.map(f)
       }
 
+  }
+
+  trait Config extends Instances {
+
+    import cats.implicits._
+    import freestyle.implicits._
+    import freestyle.config.implicits._
+
     implicit val routeGuideServiceHandler: RouteGuideService.Handler[Future] =
-      new RouteGuideHandler
+      new RouteGuideServiceHandler
 
     val grpcConfigs: List[GrpcConfig] = List(
       AddService(RouteGuideService.bindService[RouteGuideService.Op, Future])
     )
 
-    val conf: ServerW = ServerW(50051, grpcConfigs)
+    val conf: ServerW =
+      BuildServerFromConfig[ServerConfig.Op]("rpc.server.port", grpcConfigs)
+        .interpret[Try] match {
+        case Success(c) => c
+        case Failure(e) =>
+          e.printStackTrace()
+          throw new RuntimeException("Unable to load the server configuration", e)
+      }
+
+  }
+
+  trait Implicits extends CommonImplicits with Config {
+
+    import freestyle.rpc.server.handlers._
+    import freestyle.rpc.server.implicits._
 
     implicit val grpcServerHandler: GrpcServer.Op ~> Future =
       new GrpcServerHandler[Future] andThen
